+++ lib/reduced.dart
@@ -5,4 +5,5 @@ export 'src/store.dart';
 export 'src/callable.dart';
 export 'src/carrier.dart';
 export 'src/parametrized.dart';
-export 'src/functions.dart';
+export 'src/function.dart';
+export 'src/creator.dart';