+++ macos/reduced.dart
@@ -3,4 +3,4 @@
 export '../lib/src/event.dart';
 export '../lib/src/store.dart';
 export '../lib/src/callable.dart';
-export '../lib/src/functions.dart';
+export '../lib/src/function.dart';